requirements.txt
PyYAML==3.10
boto==2.27.0
